Stateless compiler (#79)
